# Speeding up vrf computation by directly producing output without proof